Make preprocessing static-only
Summary:
An annoyance in implementing the SourceCodeApi abstraction
to decouple the environment from AstEnvironment was the
realization that, given the place in the control flow we were
handling special-case dependency tracking of WildcardImports,
it was not possible to make the `processed_source_of_qualifier`
function depend only on `raw_source_of_qualifier`.

But the problem is really just that the logic was more complex
than necessary. We were deciding inside the implementation
of `processed_source_of_qualifier` whether to call a tracked
or untracked version `raw_source_of_qualifier`, and that meant
we couldn't define the abstract api entirely in terms of
raw sources.

But we can get the same behavior by instead unconditionally
filtering redundant `WildcardExport` dependencies directly in
`raw_source_of_qualifier`, at which point there's no need
to special-case anything in `processed_source_of_qualifier`.

Reviewed By: grievejia

Differential Revision: D53158275

fbshipit-source-id: a098adf99e95553ac0904765a7d4b568c07236cd