Inject processed_source_of_qualifier to SCA
Summary:
I made a mistake in my design of `SourceCodeApi`. The types all
lined up, but there are a lot of assumptions we make that are not
type-safe.

The specific problem I hit is this: SourceCodeApi hides dependency
tracking, which is something we very much want. But we special case
the use of `raw_source_of_qualifier` from `processed_source_of_qualifier`
to bypass dependency tracking. That means it is not possible to implement
`processed_source_of_qualifier` in terms of `raw_source_of_qualifier`
and get the desired dependency tracking, even though from a pure logic
perspective it's a pure function.

The fix is to inject `processed_source_of_qualifier` alongside
`raw_source_of_qualifier`, so that we can inject the special-cased dependency
tracking from the layer that controls dependencies.

--------------

Just a note about the safety of previous diffs: the problem I'm fixing
doesn't actually show up yet as of this commit, because we are still
*directly* using the AstEnvironment `get_processed_source` inside
UGE, and the AstEnvironment implementation (as opposed to the
SourceCodeApi implementation prior to this commit) did in fact special
case the dependency tracking.

But I was working on stripping away the direct AstEnvironment.ReadOnly
calls from UGE, at which point I ran into the problem.

-------------

**Does it have to be this messy?**

No. At this point in the stack I wanted to focus on building SourceCodeApi to correctly model existing Pyre logic, which requires this change.

But at the end of this effort I was able to modify internals of dependency tracking
to back this out (by pushing the special-cased dependency logic down into the raw raw parsing layer) in D53158275.

Reviewed By: kinto0

Differential Revision: D53137978

fbshipit-source-id: 8edd6a77170a4bc28843c50c081cbb4c53ff6a7f