Kill AstEnvironment.ReadOnly
Summary:
This is a big step in the process of making the entire
incremental API abstract, and one that many of the previous
refactors (especially flipping the way overlays are created)
was inching toward: AstEnvironment.ReadOnly no longer exists
except as an implementation adding extra (private) helpers
to construct SourceCodeIncrementalApi.ReadOnly values.

In the process I mark the load and store functions as assuming
an AstEnvironment - they directly depend on shared memory, so
they would not work using an alternate implementation.

Reviewed By: grievejia

Differential Revision: D53137972

fbshipit-source-id: 41c73c13433ab4d1cd4a50d9c3a17647a72bba65