Renames
Summary:
The names `raw_source` and `processed_source` have always been
confusing, and the thing is that they actually don't describe the
types very well anyway.

There's actually no use of a raw source in Pyre anyway except
to get a processed source (errors environment will grab the result,
but it only cares about the parse failure case, not the success!).

Renaming to `parse_result_of_qualfier` and just plain old
`source_of_qualifier` makes things clearer and less repetitive.

---------

Some specific notes on the state of the API relative to review suggestions in D53137975:

As of this diff almost all of the problems are addressed - there is only one module-finding function that has to be injected (`module_lookup_of_qualifier`) and source-loading function  (`parse_result_of_qualifier`). The names clearly indicate what each function does.

The only suggestion *not* followed was to link these two, either via a thunk or by forcing a `ModulePath.t` as input to the parse function. I'd be fine with that refactor in principle, but I don't think it's worth the time because:
- It won't map well to the existing dependency-tracked world, the underlying parse result table is keyed on qualifier rather than module path (and has to be for dependency tracking to work correctly)
- I don't think a bug where these are mismatched is likely in real implementations, so forcing it via types doesn't seem high-priority to me
- In an incremental scenario, types don't actually protect us anyway. If data lives beyond the lifetime of one update, you could absolutely use a no-longer valid ModulePath to try to parse.

Reviewed By: kinto0

Differential Revision: D53158273

fbshipit-source-id: 39dbdd5f0bc691189231e7def8172258d773546a