Flip Overlay creation to live on Base
Summary:
It turns out that the SourceCodeIncrementalApi refactor is actually
harder than I thought; issues start popping up in the details of the
type signatures.

One of the problems I ran into is that the current implementation of
Overlay, which is layered into two parts (AstEnvironment and ModuleTracker),
cannot be expressed as it currently exists via an abstract API. The
problem is that if an abstract `ReadOnly` doesn't know about the underlying
ModuleTracker, then there's no way that the abstract `Overlay` module
could take as input an abstract `ReadOnly` and produce the behavior
because it cannot get at the underlying module tracker via an abstract
value.

The solution I found is to move the construction of overlays out of
the Overlay module and into Base. Since an abstract Base will be
built out of closures that have direct access to the underlying
implementation, the problem goes away.

---------

This itself runs into another issue, which is that there was no good
way in tests to drill down the ReadWrite environments. I remember I
intentionally made this not possible to do *easily* when I was refactoring,
because we used to do it often but it is unsafe in general; if you grab
an "inner" read-write environment and update it directly you break
dependency tracking and the behavior is effectively undefined.

We already had an `Unsafe` module to work around this in a couple of places
where it was contextually okay; I had to extend that a bit so that
unit tests needing to construct overlays directly had a way to do so
(this is why there are new hooks on ErrorsEnvironment to unsafely grab
a class metadata environment or UGE out of it - we need it for tests).

In the process I also moved the exposed unsafe api to get a read-write
TypeEnvironment from an ErrorsEnvironment into the ErrorsEnvironment.unsafe
module; it was previously at the top-level but it's incorrect to use it in
a dependency-tracked setting so it really should be marked as Unsafe.

Reviewed By: grievejia

Differential Revision: D53137976

fbshipit-source-id: bce1c3ee0d4772f050b980a8dbe96f6ac7423592