Abstract SourceCodeIncrementalApi.Base as a record
Summary:
This commit finishes decoupling UGE's implementation from
AstEnvronment; the only coupling points now are:
- the functions explicitly marked as AssumeAstEnvironment
- the preexisting create inteface, which eventually can be modified in any of
	several ways (taking a callback as input would be my preference)

---------------

This commit completes the introduction of SourceCodeIncrementalApi,
at this point the only work needed to introduce a new implementation
(aside from the implementation itself) is to decide how the `create`
function should work, which is a relatively simple change that I would
defer until there's an actual use case.

At this point SourceCodeApi is a faithful reflection of how Pyre interfaces
worked prior to its introduction. At a few points in this process (most
notably review comments of D53137974) we've proposed changing the type
signatures to reduce duplication and the implied invariants.

In the next few diffs I'll be aiming to simplify the interfaces
by changing type signatures, tweaking how dependency tracking is dealt
with to reduce edge cases, and such.

Reviewed By: grievejia

Differential Revision: D53158272

fbshipit-source-id: 64037f6cb17492d3322e07919c1ee656eaf936c5