Move track_dependencies check down to AstEnvironment
Summary:
It is easier to keep all the dependency-tracking related edge
cases in preprocessed-source fetching in one place. We *have*
to put some of it in AstEnvironment in order to preserve the special
don't-track-redundant-dependencies logic we have today, so let's
also push the flag check there.

(Just as a recap of why the flag is here at all, this flag check is global, and
it's there because we save a ton of storage and data shuffling cost in
non-incremental contexts like `pyre check` and `pysa` by *always* disabling
tracking via a flag check.)

Reviewed By: kinto0

Differential Revision: D53137975

fbshipit-source-id: 46b1d64b859bf3d8827abe6c7e77e2e76296eb8a