Pull out helper function in AstEnvironment.ReadOnly
Summary:
We probably won't actually be modifying this implementation
much going forward, but before I wash my hands of it I did think that
it would be more readable if the bare-bones "get me a SourceCodeApi.t"
logic were split from the construction of a read-only interface with
untracked and tracked source code api fetchers.

Readability will matter if anyone has to look into dependency tracking
bugs in the future.

Reviewed By: grievejia

Differential Revision: D53140994

fbshipit-source-id: 0bdba41e80bbba9f3fbb5f6c83516d8de7d04242