Add 'exclude' field to Pyre Upgrade Configuration class
Summary:
>[This diff] is a diff that adds an 'exclude' field to the Pyre Upgrade Configuration class. The Pyre Upgrade Configuration class is used to configure how Pyre, a static type checker for Python, should be upgraded. The 'exclude' field allows users to specify a list of paths that should be excluded from the upgrade process. This can be useful if there are certain parts of the codebase that should not be upgraded or if there are known issues with specific files or directories.
The diff also includes a test plan that runs the pyre command to ensure that the changes made in the diff do not break any existing functionality. The owner of the diff is Conner Nilsen.
This diff is important because it provides a way for users to exclude certain parts of their codebase from the Pyre upgrade process. This can be useful if there are known issues with specific files or directories, or if there are certain parts of the codebase that should not be upgraded. By providing this option, users can have more control over the upgrade process and avoid potential issues.
In addition, the test plan included in the diff ensures that the changes made do not break any existing functionality. This is important because it helps to ensure that the Pyre upgrade process is reliable and that users can trust the results of the upgrade.
Overall, this diff is a valuable addition to the Pyre Upgrade Configuration class and will provide users with more flexibility and control over the upgrade process.

-- From my dear friend and colleague, Metamate

save_me_metamate

Reviewed By: grievejia

Differential Revision: D53288757

fbshipit-source-id: b70b02c39714bf54808d13e369f2cc7e4322a16f