Update the format of callee ports
Summary:
Pysa produces a taint output json file that describes issues and their frames.
In these files, each frame has ports which indicates where the taint comes from or goes.
These ports use the form `formal(x)`, `result`, etc. Those also show in the trace UI.

Internally, we actually have ports with extra information, such as the position of the argument for positional arguments, excluded names for keyword arguments, etc.
This can lead to bugs, for instance when the port appears fine in the output json (e.g, `formal(x)`) but is actually on the wrong parameter position. Another problem is having 2 ports with the same parameter name but different positions. This could lead to exporting multiple frames with the same port and kind to SAPP, which might lead to unexpected behaviors (although we never noticed it).

To prevent these issues, I propose to add the extra information within the exported port. My proposal is:
* `formal(x)` for positional argument becomes: `formal(x, position=y)` (note that positions start at 0, not 1)
* `formal(__x)` for positional-only arguments becomes: `formal(__x, position=y, positional_only)` (note that starting with `__` is not enforced, so marking it `positional_only` is useful)
* `result` is unchanged
* `formal(*restX)` becomes `formal(*args, position=X)` (mostly for clarify)
* `formal(**kw)` becomes `formal(**kwargs)` (for clarify)
* `formal(**kw)` with excluded parameters becomes `formal(**kwargs, excluded=[a, b, c])`

Reviewed By: alexkassil

Differential Revision: D53224284

fbshipit-source-id: 153a04a0928ed42ef9e2513ee8933fce22ce5f74