Remove daemon_failer.py and cleanup daemon_querier_test.py and daemon_querier.py
Summary:
VSCode's language client must send a request for every file under the source root. For example, fbcode root is as follows:
fbcode root
project 1
project 2
For typechecking, this is usually not an issue. Project 1 and project 2 are each indipendent source roots and have no overlap.
With code navigation, the source root is at fbcode root. We could also have a local root in project 1 (for example instagram-server). This means we spin up two code navigation servers and both get all requests. D47655034 specifies a syntax for fast-failing one one of them.

The fast-fail logic is not preferred for a number of reasons:
VSCode waits on both language servers before showing any results: if either are slow for any reason, there is user impact
We keep both servers up using twice as many resources as we would with one server
Metrics: even if we prevent logging on our side, vscode chronicle events will log two events (one for each language client <> language server). This is confusing and may mess up dashboards

As part of the effort to amend the above, we remove DaemonQueryFailer

Reviewed By: grievejia

Differential Revision: D53454746

fbshipit-source-id: 8c64f78375fd96dca119fde7de518c95275182e4