Make it clear where ScratchProject relies on OUnit2
Summary:
Jia and I were chatting about devx today and the question of whether
we can decouple ScratchProject from OUnit2 came up twice:
- I mentioned trying out alcotest, an alternative to OUnit2 that I think is a
  little better (not enough better to rewrite tests, I don't think I'll
  actually follow up on this), or potentially jane street's ppx
  test libraries (probably not a good idea for us for several reasons,
  although they are nice and I'd definitely use them on a side project!)
- Jia mentioned that ideally he'd like to have a fast iteration loop
  where you could paste a code sample directly into `dune utop`

In both cases, a signficant issue is that ScratchProject is coupled to OUnit2.

I'm not going to decouple it right now because I have other priorities, but I
wanted to be able to see at a glance what all the places we would need to update
are, so I deleted the `open OUnit2` in `test.ml`, and I figured we might as
well commit this change.

Reviewed By: kinto0

Differential Revision: D53593369

fbshipit-source-id: dc8bda57af22dac460ea61c408c582396d6cd517