[easy Pass context as last argument, positionally
Summary:
I would like to takle the bulk refactor that Pradeep poked at in
D48926926 - our test cases need to run in isolation, this is one of
the most severe usability problems when changing the type checker.

To unblock that refactor, it will be extremely helpful for the helper
functions to take context as the last positional argument, because that
will enable us to trivially curry the test functions which is what the framework
wants.

This diff makes that change in `test.ml` and `integrationTest.ml`;
I do not modify how any of the downstream tests actually run in this
diff, I just turn the crank to make them work with the new signature.

Reviewed By: kinto0

Differential Revision: D53597429

fbshipit-source-id: ad28ca231f3265bf4d5d968264f6de985aa6258f