Isolate tests in annotationTest.ml
Summary:
At this point I am finished with refactoring tests for their own
sake, I'll leave modules I don't actually need to change urgently
for the future (to fix whenever they get in my way) or for BE days.

But the immediate reason I wanted to get test isolation in the first
place is because I decided to change the behavior of `reveal_type`,
which currently hides type errors and can lead to confusion
in sandbox results.

This refactor took me forever because it was extremely difficult
to actually find and fix the relevant unit tests.

So, I'm going to perform the refactor on each of the test modules
where changing `reveal_type` led to test failures; there are
just a small handful of them but they do tend to be some of the
bigger ones.

Reviewed By: kinto0

Differential Revision: D53623455

fbshipit-source-id: e4daa4daf2e7694c23b3c59bed34505e3cb99d24