Isolate test cases in callableTest.ml
Summary:
I wanted to very carefully do a mid-size refactor and record the precise
steps I took, so that it will be easy to repeat on other modules. We might
even be able to make a python script to do this, although I'm skeptical
that it's worth the effort (various edge cases call for manual intervention,
so using editor macros might be a better choice).

The conceptual I used for this:
- Indent everything so python snippets are at 14 spaces (usually that means
  add 8 spaces because most of our samples are currently indented 6)
- Convert the start of every test function into `= test_list [`
- Convert the end of every test function into `]`, typically deleting a
  trailing `()`
- Delete all instances of `context` being passed explicitly
- Search for lines starting with `assert_.*` and (this does require
  a little care because some places have to be skipped, especially
  asserts in the Python samples!) convert that to
  `labeled_test_case __FUNCTION__ __LINE__ @@ assert_.*`

------------------------

This was pretty easy to do with vim shortcuts:
- You can use line selection and `I` to insert 8 spaces in the entire file
- You can search for `^let ` in order to find the start of each test function,
  which makes it pretty easily to manually get the start and end of the `test_list`
- Either a search and replace or a search and repeat editor action works for
  the assertion.

The following three search-and-replace commands are particularly helpful:
```
# change most of the test headers
:%s/context =/= test_list [/g

# change the assertion functions to wrap them in labeled_test_case
:%s/(assert_[a-z_]*_errors)/labeled_test_case __FUNCTION__ __LINE__ @@ \1/g

# delete remaining places where context is passed explicitly
:%/s/context//g
```
They won't do everything for you by any means, but they'll take care of the bulk of the changes on many integration test files.

Reviewed By: kinto0

Differential Revision: D53623456

fbshipit-source-id: 847c513182969ba5474230031d8729a8a262960b