Push UGE creation into a CreateHandle
Summary:
This commit makes it more obvious how to customize creation of an
ErrorsEnvironment that is *not* backed by an AstEnvironment, by making the
`create` function originally defined in UGE and propagated by the
`Environment` functor work with a `CreateHandle`, that is a pair
of `(SourceCodeIncrementalApi.Base.t, AstEnvironment.t option)`.

Existing constructions are easily handled by having a convenience function
`create_with_ast_environment` in the `ErrorsEnvironment` top-level, but now it
should be easy for a new entrypoint to create an environment based off of some
other `SourceCodeIncrementalApi.Base.t` implementation; in that case just pass
`None` for the `AstEnvironment.t`, which is only in the signature because some
functions we won't need to support in buck-calls-pyre (especially saved state
load and save) require an actual concrete `AstEnvironment`.

Reviewed By: kinto0

Differential Revision: D53669254

fbshipit-source-id: 5dea79fdce479ca0cafab6e04c4ad9541f4e85fd