]Add get_document_symbols methods to the codebase
Summary:
Context:
We want to turn on the document symbol feature on for code navigation and test it end to end.

Changes:
In this diff, we add the get_document_symbols to the codebase so that it can call various implementations. We will turn the feature on for code navigation and test it end to end in the next diff in this stack. We also add testing infrastructure in case it's useful for investigating performance and for code maintainability reasons.

Reviewed By: kinto0

Differential Revision: D53635234

fbshipit-source-id: 1acc9e388369e03498fed00b96f39edef6fe1534