Add document symbol response
Summary: This diff focuses on adding the data structures for a document symbol response. The structures are drawn from here https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#documentSymbol.I used the existing Ast.Location class to represent Range which is in the LSP protocol. The rest of the fields are the same as the LSP protocol.

Reviewed By: kinto0

Differential Revision: D53594477

fbshipit-source-id: 91c098b86803fc53078d207324710d04cf54e3cf