Isolate tests in typeVariableTest
Summary:
This is another module where tests broke when I tried not
hiding errors in `reveal_type`, so I want to refactor it
before I ship that change.

The only thing that wasn't a pretty simple search-and-replace
type tweak in this diff was the one test where we need to
reset our recursive variable namespaec for determinism; I
had to wrap the assert helper in a closure so that I could
put the side-effecting operation in front.

Reviewed By: connernilsen

Differential Revision: D53634454

fbshipit-source-id: 9a778e06beadb75be46569cd000a44c98ae1aaa8