Apply mode to files even if global mode matches
Summary:
With the configurationless migration and Global Lazy type checking, I'd like to add `# pyre-strict` comments to files that are covered by Pyre. The goal here will *not* be to add these headers to all Python files, and we will eventually remove this redundant header once Per-Target type checking takes over or Buck-powered Pyre settings become more integrated with the Pyre Language Server.

The reason I would like to do this is to run Global Lazy in a default `ignore-all-errors` mode and let the Pyre mode header determine if type errors should be returned.

Reviewed By: MaggieMoss

Differential Revision: D53714213

fbshipit-source-id: 17142332f923866699f5dfb85d36c641de27ca74