`reveal_type` should not hide type errors
Summary:
I recently was playing with sandbox examples, and incorrectly filed a bug
because I didn't realize that our implementation of `reveal_type` was swallowing
any type errors produced when handling the inner expression.

I believe this was an unintended side effect of removing errors from `State.t`
in the stack starting with D20370377.

In my opinion this is probably not desirable behavior, I doubt I'm the only
person who would be confused by this.

Reviewed By: samwgoldman

Differential Revision: D53539722

fbshipit-source-id: 7ae062ca0eda188ae638f94b988ef63f35657930