Propagate resolution in reveal_type [1]: new test case
Summary:
It was noted in review of an earlier change where I stopped hiding type
errors inside of `reveal_type` that we ought to be propagating the
resolution, which could potentially contain variables that should
be in scope.

I prefer to write the test case illustrating an issue in a separate
commit from the fix so that we can see the effect of the fix
in a diff, hence splitting this bugfix into two commits.

Reviewed By: samwgoldman

Differential Revision: D53736215

fbshipit-source-id: afd8073c8d234d4650b9738781538449b3818ebb