Isolate some ModuleTrackerTest cases
Summary:
I wanted to try isolating test cases in a non-integration test
suite, and this one came to mind because I've spent so much time
in it over the past two years :)

This diff proved to me that it's typically not very hard to
convert non-integration tests to use labeled cases, although it
does often involve some manual work to make the helper functions
take context as the last positional argument.

I didn't conver *all* of the test cases because a few of them
are relying on a bulk setup to run a bunch of assertions, and
I wouldn't want to pay the performance cost of isolating those
cases (which aren't that likely to break anyway).

Reviewed By: grievejia

Differential Revision: D53740127

fbshipit-source-id: cf122442af266e3a43236bf03e96a22ee4e1a046