Relax invariant in ModuleTracker
Summary:
When investigating OOMs, Kyle found clear proof that we can hit a case where
the `is_external` flag is mismatched on *insert* into ModuleTracker.

It was already known that it could differ on removal because symlink reading
will fail when Buck deletes a path, causing `is_external` to flip from `false`
to `true` just as we delete the `ModulePath`.

But the logs Kyle found demonstrated that it can happen due to data races
even on insert; it's likely that things have to be running abnormally slowly
for this to be likely (this was a server where Pyre and Buck were running
into OOMs), but there's no sense creating unnecessary errors that could
prove a distraction when debugging.

As an added benefit, the invariant is simpler, more consistent, and easier
to clearly describe in a comment that makes it clear why `is_external`
can differ.

Reviewed By: connernilsen

Differential Revision: D53816861

fbshipit-source-id: ef0fa36009f2583d1e81717ab34ee06769c53091