Use Python 3.12 for the OSS playground
Summary:
In response to Shantanu's request in
https://github.com/facebook/pyre-check/issues/816
plus general hygeine, I am redeploying the playground on a more
recent AMI and using Python 3.12.

This commit modifies the setup script to get 3.12 (which is not
the default Python version for Ubuntu) and use it; this recipe
should be pretty maintainable because the `deadsnakes` ppa maintains
a good-size set of python minor versions (both newer and older
than the main Ubuntu repositories do).

Reviewed By: connernilsen

Differential Revision: D53782860

fbshipit-source-id: 83d03e5c9901ad8fe6624fb1f3b746c0d39d08e8