Support typing.reveal_type [2] the fix
Summary:
At a recent PyCon typing summit we decided to add `typing.reveal_type` to
the standard library, and have it print types at runtime so that users
could easily
- mark code with `reveal_type` to get type checker signal, and *also*
- run the code without crashing (the older `reveal_type` behavior relied
  on a function that didn't exist at runtime, so it would crash) and even
  get some feedback about the runtime types.

Pyre is the only type checker that doesn't support this according to
https://htmlpreview.github.io/?https%3A%2F%2Fraw.githubusercontent.com%2Fpython%2Ftyping%2Fmain%2Fconformance%2Fresults%2Fresults.html

Let's fix it!

This commit follows up on the previous one (which set up a test) with the actual fix.

Reviewed By: grievejia

Differential Revision: D53771942

fbshipit-source-id: 4501245b7236d724db6f5c3de7885c429e0567a7