reveal_type passes arg out [1]: test current behavior
Summary:
I just realized when looking at `typing.reveal_type` that Pyre's existing
implementation of `reveal_type` was suboptimal: Pyre is often bad at type
inference of expressions, so it's valuable to be able to use reveal_type
inside a complex expression (this will be much more true if we use
bidirectional inference in the future!).

But the existing implementation assumes `reveal_type` is `None`, i.e.
equivalent to the runtime behavior of just printing something.

This isn't ideal, but it's more explicitly wrong for `typing.reveal_type`,
which has actual runtime behavior of:
- print the runtime type
- return the value, so that it can be used inside a bigger expression

As a result, I want to fix our special-cased branch to correctly reflect
the actual behavior (and the stub) by propagating the revealed type.
-----

This commit adds a test illustrating the current behavior, and we'll
see the behavior change in the next diff.

Reviewed By: samwgoldman

Differential Revision: D53825253

fbshipit-source-id: 1702f764bbc0bb6d80a7790a87695b9524592349