Implement qualification for `assert_type`
Summary:
We need to qualify the second argument of `assert_type` under the assumption
that it is a type, similar to existing qualifcation on `cast` calls.

This is needed to support forward references, and also any non-builtin
type (the conformance tests use mostly builtin types, but in real code
`assert_type` would actually usually fail without qualification support).

Reviewed By: grievejia

Differential Revision: D53852673

fbshipit-source-id: 978fc3b13791f6b715190b25f9fd04e084074114