Implement basic support for `typing.assert_type`
Summary:
This commit adds basic, best-effort support based on `typeCheck.ml`
handling only, using `Type.equal` to assert that the inferred type
matches the asserted type exactly.

**What remains to be done in later diffs**

- There are two TODOs on this diff, both of which require preprocessing changes
   that I split into stacked diffs:
   - support forward references, which requires qualification
   - support `|` syntax, which requires the union shorthand preprocessing step
- It's not very urgent based on current conformance tests, but we probably want
  to eventually normalize union ordering.
  - We potentially would also want to use type ordering on protocols, although I
    doubt this will ever really be necessary for conformance tests.

In addition, we'll have to figure out how to handle `assert_type` if
we ever move away from eager inference of expression types in our
type checking algorithm.

**What are the open questions for review?**

What's *not* an open question is whether type equality makes sense (as opposed
to some kind of two-way subtyping test). The `assert_type`'s type-equality
semantics aren't popular on the team, see
[this discussion](https://discuss.python.org/t/is-assert-type-checking-for-exact-match-or-a-subtype-relationship/45299/17),
but there's really no alternative given the current state of open source.

So I think the biggest open question is
whether we should create a a new error code; I use
`IncompatibleParameterType` because it works (the message is clear, the spec
doesn't care, and we already use this error for some other special-cased
functions so there's no precedent of avoiding it in non-vanilla cases), but it
would be easy enough to add an `AssertTypeFailure` - probably identical to
`IncompatibleParameterType` except for the error code - if we think that's
better.

Reviewed By: kinto0

Differential Revision: D53852502

fbshipit-source-id: d2e83d88147515130d3ee5ce89cace16391fbfc8