Define kind coverage (i.e., sources, sinks, and transforms) that is used for computing rule coverage
Summary:
The next step is to compute the category / rule coverage, which however depends
on defining the sources, sinks, and transforms that are used in such
computation, which is done by this diff. For example, not all sinks (e.g.,
`ExtraTraceSink`) are interesting from the perspective of computing category
coverage.

In particular, sanitizers are treated as taint transforms in the
implementation, but they shouldn't be considered as user-defined taint or taint
transforms, since their semantics is to prevent the propagation of certain
kinds of taint, instead of introducing / defining new kinds of taint or taint
transforms.

This diff also implements the logic of extracting the sources, sinks, and
transforms (that are used for computing coverage metrics) from models.

Reviewed By: arthaud

Differential Revision: D53813348

fbshipit-source-id: 6d4e0c4c4b05e114de5f4cc5679d9286f9988c5a