Define whether a rule is considered as being "covered"
Summary:
By consensus, a rule is considered as being "covered", if
- at least one of its sources is used in some model,
- at least one of its sinks is used in some model, and
- at least one of its transforms (if any) are "used" in some model.

The above definition also applies to multi-source rules, which in
implementation is translated into two sub-rules (that share the same rule
code). A multi-source rule is special in the sense that, both of its sub-rules
need to be covered, in order for the multi-source rule to be considered as
being "covered".

Reviewed By: arthaud

Differential Revision: D53824554

fbshipit-source-id: 3b8f9ba3f33710b8631c08c44e04e4d538354f3c