Test some additional assert_type edge cases
Summary:
It appears that we handle union ordering and legacy containers
like Tuple / tuple out of the box.

But our Callable and BoundMethod types do not work out of the box,
we'll need a fix for these (in the long run, we probably need to remove this
information - used by Pysa - from Type.t and put it somewhere else; we probably
could store it in the resolution similar to how we store extra data for
refinement, although it would be ideal to have a separate pass and get this
entirely out of type checking).

This diff simply records the current behaviors, I'll work on a fix for the callable
and method types later.

Reviewed By: grievejia

Differential Revision: D53923070

fbshipit-source-id: d5b606881fe4695b809c2a8f239ada73596f3b0a