Support BoundMethod in assert_type equivalence
Summary:
This diff extends the canonicalization transform in
`equivalent_for_assert_type`` to "unpack" bound methods. Bound methods occur in
Pyre's representation of the type of regular and class methods, but *not*
static methods, and the underlying callable signatures are always "full", in
the sense that they include `cls` or `self` in the arguments list.

So to unpack them and canonicalize the types, we need to both extract the inner
callable, and strip the first argument (which can be named or positional, and
the name is technically not relevant) from all signatures, both implementation
and any overloads.

As far as I know thus far, this will bring Pyre's assert_type in line with the
minimal functionality required by conformance tests; there do exist more complex
cases like structural type comparison (where it is arguable what is required but
discussion has favored using structural rather than nominal equality) but the
conformance tests currently avoid those cases.

Reviewed By: grievejia

Differential Revision: D53923071

fbshipit-source-id: 74eac2c86f0be82c68c19c177d36853851ed4f6b