Add migration_status key to set of valid Pyre configuration keys
Summary:
In the configurationless script, I add a "migration_status" key to the .pyre_configuration.local files that are migrated. MaggieMoss and I originally discussed adding this key for 2 reasons:
1. We could use this to slowly roll out the Configurationless migration to one project at a time, starting with projects we determine are the most important to migrate first.
2. It could be useful for future migrations to determine what state the given project is in, and future configurationless migration logic could be based around this migration status.

As of right now, I'm in favor of starting to roll this out to as many projects as possible (after a small test with around 10-15 starter projects), so the migration status key might not be as useful for that reason, but it could still be used for future migrations. I'm curious to hear others' opinions on this, but I think this should be a relatively safe migration to do, so I'd like to go for it.

Alternatively, I could go in and filter out configurations we run in the regular 'daily' migration by looking for this configuration option, and then we can enable a migration on a project by manually going in and adding the option if we'd like to do a slower rollout.

Let me know what you think.

insanityrfc
👀

Reviewed By: grievejia

Differential Revision: D54044374

fbshipit-source-id: b9e7aa1784f7c56cd2732efa54725a06a0ed9e50