Fix script to handle ignore paths from root
Summary:
I wasn't aware that we allow the `//` in something like `//path/to/my/file.py` to act as a reference to the root of the global `.pyre_configuration`. Since this is used in a good amount of configurations we're planning to upgrade, this diff adds support for projects that reference the project root or use relative paths in their `ignore_all_errors` path prefix lists.

ros_navigation_path_planning

Reviewed By: ljw1004

Differential Revision: D54135305

fbshipit-source-id: 505c8939e9be797ea961febe1b92d5126686cf35