Add a test to clarify the meaning of the decorator fully qualified name matching clause
Summary:
The meaning of decorator fully qualified name matching clause is a bit
confusing to the users. Its actual meaning is to match against the fully
qualified names of the identifiers that are local to a file (e.g., local
variables) who point to (or hold a reference to) the decorator, as opposed to
matching against the fully qualified name of the decorator itself.  That is,
the fully qualified name of a callable.

Reviewed By: arthaud

Differential Revision: D54122123

fbshipit-source-id: 96eb6a18ae2c8097ff244a7a55d8cff385973ecd