Change return type of document symbol response to a list of document symbol
Summary:
Context:
We are working on document symbol in daemon_querier.py. This is in preparation for the OCaml <> Python interface, in an effort to support document symbol in the backend.

Goal:
Depricating Pyls for document symbols

Changes:
As per the discussion in this diff D53963996, the return type of document symbols should be a list of DocumentSymbol rather than a single DocumentSymbol

Reviewed By: kinto0

Differential Revision: D54006451

fbshipit-source-id: d773f974ffaef9ce26e458ac4ca0e664d9af91e2