Rename document symbol response
Summary:
Context:
We are working on document symbol in daemon_querier.py. This is in preparation for the OCaml <> Python interface, in an effort to support document symbol in the backend.

Goal:
Depricating Pyls for document symbols

Changes:
We improve variable renaming throughout the entire document symbol implementation (including the existing Python implementation)

Reviewed By: kinto0

Differential Revision: D54023332

fbshipit-source-id: 174408a30cce235909a0c8464570c5be69941a89