Use a custom serialization method for document symbol
Summary:
Context:

As part of rolling out document symbol, we consider the performance. After conducting an investigation and rolling out document symbol to 50% of fbcode users, we observe that document symbol is fast for most cases except for p97 and beyond https://fburl.com/scuba/vscode_chronicle_language_server_events/vsrzb10d

Separately, we have conducted this investigation https://fb.workplace.com/groups/657557118131740/permalink/1446334929253951/

and showed that serialization presents a bottleneck for document symbol.

In this diff, we want to understand the effects of a custom serialization method which improves the overhead based on preliminary results.

Reviewed By: connernilsen

Differential Revision: D53734720

fbshipit-source-id: 7ceb534a4e062a27656e9d9928052c6ce333e3cb